move docstring tests under options tests and get them to pass
