+++ test/default_style.jl
@@ -1372,143 +1372,6 @@
 
     end
 
-    @testset "docs" begin
-        str = """
-        \"""
-        doc
-        \"""
-        function f()
-            20
-        end"""
-        t = run_pretty(str, 80)
-        @test length(t) == 12
-
-        normalized = """
-        \"""
-        doc
-        \"""
-        function f()
-            20
-        end"""
-
-        str = """
-        \"""doc
-        \"""
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        str = """
-        \"""
-        doc\"""
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        str = """
-        \"""doc\"""
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        str = """
-        "doc
-        "
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        str = """
-        "
-        doc"
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        str = """
-        "doc"
-        function f()
-            20
-        end"""
-        @test fmt(str) == normalized
-
-        # test aligning to function identation
-        str_ = """
-            "doc"
-        function f()
-            20
-        end"""
-        str = """
-        "doc"
-        function f()
-            20
-        end"""
-        @test fmt(str_) == normalized
-
-        str = """\"""
-                 doc for Foo
-                 \"""
-                 Foo"""
-        @test fmt(str) == str
-        t = run_pretty(str, 80)
-        @test length(t) == 11
-
-        str = """
-        \"""
-        doc
-        \"""
-        function f()    #  comment
-            20
-        end"""
-        @test fmt(str) == str
-
-        #
-        # Issue 157
-        str = raw"""
-        @doc \"""
-           foo()
-        \"""
-        foo() = bar()"""
-        @test fmt(str) == str
-
-        str = raw"""
-        @doc doc\"""
-           foo()
-        \"""
-        foo() = bar()"""
-        @test fmt(str) == str
-
-        str = raw"""@doc "doc for foo" foo"""
-        @test fmt(str) == str
-
-        str = raw"""@doc \"""doc for foo\""" foo"""
-        @test fmt(str) == str
-
-        str = raw"""@doc doc\"""doc for foo\""" foo()"""
-        @test fmt(str) == str
-
-        str = raw"""@doc foo"""
-        @test fmt(str) == str
-
-        # issue 160
-        str = """
-        module MyModule
-
-        import Markdown: @doc_str
-
-        @doc doc\"""
-            foo()
-        \"""
-        foo() = bar()
-
-        end # module"""
-        @test fmt(str) == str
-    end
 
     @testset "strings" begin
         str = """
@@ -4804,151 +4667,6 @@ some_function(
         @test fmt(str_, 4, 1) == str
     end
 
-    @testset "Format docstrings" begin
-        unformatted = """
-        \"""
-        This is a docstring
-
-        ```@example
-        a =  1
-         b  = 2
-         a + b
-        ```
-
-        ```jldoctest
-        a =  1
-        b  = 2
-        a + b
-
-        # output
-
-        3
-        ```
-
-        ```jldoctest
-        julia> a =  1
-        1
-
-        julia> b  = 2;
-
-        julia>  a + b
-        3
-
-        julia> function test(x)
-               x + 1
-               x + 2
-               end;
-        ```
-        \"""
-        function test(x) x end"""
-
-        formatted = """
-        \"""
-        This is a docstring
-
-        ```@example
-        a = 1
-        b = 2
-        a + b
-        ```
-
-        ```jldoctest
-        a = 1
-        b = 2
-        a + b
-
-        # output
-
-        3
-        ```
-
-        ```jldoctest
-        julia> a = 1
-        1
-
-        julia> b = 2;
-
-
-        julia> a + b
-        3
-
-        julia> function test(x)
-                   x + 1
-                   x + 2
-               end;
-
-        ```
-        \"""
-        function test(x)
-            x
-        end"""
-        @test fmt(unformatted) == formatted
-    end
-    @testset "Multi-line indented code-blocks" begin
-        unformatted = """
-        \"""
-            fmt(
-            )
-        \"""
-        function fmt() end"""
-
-        formatted = """
-        \"""
-            fmt(
-            )
-        \"""
-        function fmt() end"""
-        @test fmt(unformatted) == formatted
-    end
-    @testset "Empty line in docstring" begin
-        unformatted = """
-        \"""
-
-        \"""
-        function test() end"""
-
-        formatted = """
-        \"""
-        \"""
-        function test() end"""
-        @test fmt(unformatted) == formatted
-    end
-    @testset "Indented docstring" begin
-        unformatted = """
-        begin
-            \"""
-            Indented docstring
-
-            with multiple paragraphs
-            \"""
-            indented_item
-        end"""
-        formatted = """
-        begin
-            \"""
-            Indented docstring
-
-            with multiple paragraphs
-            \"""
-            indented_item
-        end"""
-        @test fmt(unformatted) == formatted
-        short = """
-        begin
-            "Short docstring"
-            item
-        end
-        """
-        short_formatted = """
-        begin
-            \"""
-            Short docstring
-            \"""
-            item
-        end
-        """
-        @test fmt(short) == short_formatted
-    end
 
     @testset "issue 260 - BracesCat" begin
         str = "{1; 2; 3}"
